requirements.txt
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.2 